header files modified for double precision program
git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/HiRep@31 37a69121-b030-0410-a54a-fdca83ffee47

+++ Include/dirac.h
@@ -9,8 +9,8 @@ typedef enum {
 } block_selector;
 
 void Dphi_(block_selector B, suNf_spinor *out, suNf_spinor *in);
-void Dphi(float m0, suNf_spinor *out, suNf_spinor *in);
-void g5Dphi(float m0, suNf_spinor *out, suNf_spinor *in);
+void Dphi(double m0, suNf_spinor *out, suNf_spinor *in);
+void g5Dphi(double m0, suNf_spinor *out, suNf_spinor *in);
 
 void Dphi_dble_(block_selector B, suNf_spinor_dble *out, suNf_spinor_dble *in);
 void Dphi_dble(double m0, suNf_spinor_dble *out, suNf_spinor_dble *in);
